New Changes from Linux Segmentation works (only one dimension Grey for now) Working on the new GUI (not this one)

+++ Qt Code/utilities.h
@@ -9,6 +9,6 @@ using namespace Eigen;
  * The Utilities file contains some commom use functions.
  */
 
-double maxInfNormInNeighood(MatrixXi M, int indexRow, int indexCol);
+float maxInfNormInNeighood(MatrixXi M, int indexRow, int indexCol);
 
 #endif // UTILITIES_H
